@@ -1,7 +1,13 @@
+# frozen_string_literal: true
+
 require "application_system_test_case"
-require 'test_helper'
+require "test_helper"
 
 class ContactTest < ApplicationSystemTestCase
+  setup do
+    mock_cloudinary_upload
+  end
+
   test "Fill form - Contactez-nous" do
     marseille_site = FactoryBot.create(:site)
     SitesController.any_instance.stubs(:find_domain_from_url).returns(marseille_site.domain_name)
@@ -10,9 +16,9 @@ class ContactTest < ApplicationSystemTestCase
     visit root_path
     click_on "Contactez-nous"
     @site = marseille_site
-    fill_in('contact_user', :with => 'John')
-    fill_in('contact_email', :with => 'John@test.com')
-    fill_in('contact_body', :with => "It's a trap")
+    fill_in("contact_user", with: "John")
+    fill_in("contact_email", with: "John@test.com")
+    fill_in("contact_body", with: "It's a trap")
     click_button "Envoyer"
     assert_equal root_path, page.current_path
     # save_and_open_page
