@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 # test/application_system_test_case.rb
 require "test_helper"
 
 class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
-  driven_by :headless_chrome # Update this line
+  driven_by :selenium_headless
 end
